LPS-54903 move class from portal-imp to portal-service since it is intended to be used from other modules
