+++ modules/apps/layout/layout-admin-web/src/com/liferay/layout/admin/web/servlet/taglib/ui/BaseLayoutSetFormNavigatorEntry.java
@@ -15,10 +15,10 @@
 package com.liferay.layout.admin.web.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.servlet.taglib.ui.BaseFormNavigatorEntry;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.LayoutSet;
-import com.liferay.portal.servlet.taglib.ui.BaseFormNavigatorEntry;
 
 import java.util.Locale;
