+++ portal-impl/src/com/liferay/portal/servlet/taglib/ui/DeprecatedFormNavigatorEntry.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.servlet.taglib.ui.BaseFormNavigatorEntry;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 
 import java.util.Locale;
