+++ portal-service/src/com/liferay/portal/kernel/servlet/taglib/ui/BaseFormNavigatorEntry.java
@@ -12,12 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.portal.kernel.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.servlet.ServletContextPool;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
 import com.liferay.portal.model.User;
 import com.liferay.portal.spring.context.PortalContextLoaderListener;
